tests/wpt/tests/dom/nodes/getElementsByClassName-13.htm

<!doctype html>
<html class="a">
 <head>
  <title>element.getElementsByClassName(): adding an element</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
 </head>
 <body class="a">
  <div id="log"></div>
  <script>test(function() {
                 var collection = document.body.getElementsByClassName("a");
                 var ele = document.createElement("x-y-z");
                 ele.className = "a";
                 document.body.appendChild(ele);
                 assert_array_equals(collection, [ele]);
               })
  </script>
 </body>
</html>
